@@ -9,7 +9,7 @@
     "start": "node ./dist/app",
     "predev": "npx tsc",
     "dev": "nodemon ./dist/app",
-    "lint": "eslint .",
+    "lint": "eslint . --ext .ts",
     "format": "prettier --write '**/*.{js,ts}'",
     "pretest": "tsc -p tsconfig.test.json",
     "test": "export NODE_ENV=test && mocha --exit 'dist/tests/**/*.spec.js'"
@@ -41,13 +41,16 @@
     "network"
   ],
   "devDependencies": {
+    "@types/i18n": "^0.13.12",
     "@types/mocha": "^10.0.9",
-    "@types/node": "^20.5.0",
+    "@types/node": "^20.16.15",
     "@types/node-schedule": "^2.1.0",
     "@types/qrcode": "^1.5.2",
+    "@typescript-eslint/eslint-plugin": "^8.11.0",
+    "@typescript-eslint/parser": "^8.11.0",
     "chai": "^4.3.4",
     "chokidar": "^3.5.3",
-    "eslint": "^8.15.0",
+    "eslint": "^8.57.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
