@@ -1,26 +1,28 @@
-const { createHash, randomBytes } = require('crypto');
-const lightning = require('lightning');
-const lnd = require('./connect');
-const { logger } = require('../logger');
+import { createHash, randomBytes } from 'crypto';
+import lightning from 'lightning';
+import lnd from './connect';
+import { logger } from '../logger';
 
-const createHoldInvoice = async ({ description, amount }) => {
+const createHoldInvoice = async ({ description, amount }: { description: string, amount: number }) => {
   try {
     const randomSecret = () => randomBytes(32);
-    const sha256 = buffer => createHash('sha256').update(buffer).digest('hex');
+    const sha256 = (buffer: Buffer) => createHash('sha256').update(buffer).digest('hex');
     // We create a random secret
     const secret = randomSecret();
     const expiresAt = new Date();
     expiresAt.setSeconds(expiresAt.getSeconds() + 3600);
 
     const hash = sha256(secret);
-    const cltv_delta = parseInt(process.env.HOLD_INVOICE_CLTV_DELTA);
+    const holdInvoiceCltvDelta = process.env.HOLD_INVOICE_CLTV_DELTA;
+    // sticking to semantics of JS code as requiring HOLD_INVOICE_CLTV_DELTA to be defined breaks tests
+    const cltv_delta = holdInvoiceCltvDelta === undefined ? NaN : parseInt(holdInvoiceCltvDelta);
     const { request, id } = await lightning.createHodlInvoice({
       cltv_delta,
       lnd,
       description,
       id: hash,
       tokens: amount,
-      expires_at: expiresAt,
+      expires_at: expiresAt.toISOString(),
     });
 
     // We sent back the response hash (id) to be used on testing
@@ -30,7 +32,7 @@ const createHoldInvoice = async ({ description, amount }) => {
   }
 };
 
-const settleHoldInvoice = async ({ secret }) => {
+const settleHoldInvoice = async ({ secret }: { secret: string }) => {
   try {
     await lightning.settleHodlInvoice({ lnd, secret });
   } catch (error) {
@@ -38,7 +40,7 @@ const settleHoldInvoice = async ({ secret }) => {
   }
 };
 
-const cancelHoldInvoice = async ({ hash }) => {
+const cancelHoldInvoice = async ({ hash }: { hash: string }) => {
   try {
     await lightning.cancelHodlInvoice({ lnd, id: hash });
   } catch (error) {
@@ -46,7 +48,7 @@ const cancelHoldInvoice = async ({ hash }) => {
   }
 };
 
-const getInvoice = async ({ hash }) => {
+const getInvoice = async ({ hash }: { hash: string }) => {
   try {
     return await lightning.getInvoice({ lnd, id: hash });
   } catch (error) {
@@ -54,7 +56,7 @@ const getInvoice = async ({ hash }) => {
   }
 };
 
-module.exports = {
+export {
   createHoldInvoice,
   settleHoldInvoice,
   cancelHoldInvoice,
