@@ -1,23 +1,31 @@
-const { subscribeToInvoice } = require('lightning');
-const { Order, User } = require('../models');
-const { payToBuyer } = require('./pay_request');
-const lnd = require('./connect');
-const messages = require('../bot/messages');
-const ordersActions = require('../bot/ordersActions');
-const { getUserI18nContext, getEmojiRate, decimalRound } = require('../util');
-const { logger } = require('../logger');
+import { subscribeToInvoice } from 'lightning';
+import { Order, User } from '../models';
+import { payToBuyer } from './pay_request';
+import lnd from './connect';
+import * as messages from '../bot/messages';
+import * as ordersActions from '../bot/ordersActions';
+import { getUserI18nContext, getEmojiRate, decimalRound } from '../util';
+import { logger } from '../logger';
+import { HasTelegram } from '../bot/start';
+import { IOrder } from '../models/order';
 
-const subscribeInvoice = async (bot, id, resub) => {
+const subscribeInvoice = async (bot: HasTelegram, id: string, resub: boolean = false) => {
   try {
     const sub = subscribeToInvoice({ id, lnd });
     sub.on('invoice_updated', async invoice => {
       if (invoice.is_held && !resub) {
         const order = await Order.findOne({ hash: invoice.id });
+        if (order === null)
+          throw new Error("order was not found");
         logger.info(
           `Order ${order._id} Invoice with hash: ${id} is being held!`
         );
         const buyerUser = await User.findOne({ _id: order.buyer_id });
+        if (buyerUser === null)
+          throw new Error("buyerUser was not found");
         const sellerUser = await User.findOne({ _id: order.seller_id });
+        if (sellerUser === null)
+          throw new Error("sellerUser was not found");
         order.status = 'ACTIVE';
         // This is the i18n context we need to pass to the message
         const i18nCtxBuyer = await getUserI18nContext(buyerUser);
@@ -47,11 +55,13 @@ const subscribeInvoice = async (bot, id, resub) => {
             rate
           );
         }
-        order.invoice_held_at = Date.now();
+        order.invoice_held_at = new Date();
         order.save();
       }
       if (invoice.is_confirmed) {
         const order = await Order.findOne({ hash: id });
+        if (order === null)
+          throw new Error("order was not found");
         logger.info(
           `Order ${order._id} - Invoice with hash: ${id} was settled!`
         );
@@ -70,12 +80,16 @@ const subscribeInvoice = async (bot, id, resub) => {
   }
 };
 
-const payHoldInvoice = async (bot, order) => {
+const payHoldInvoice = async (bot: HasTelegram, order: IOrder) => {
   try {
     order.status = 'PAID_HOLD_INVOICE';
     await order.save();
     const buyerUser = await User.findOne({ _id: order.buyer_id });
+    if (buyerUser === null)
+      throw new Error("buyerUser was not found");
     const sellerUser = await User.findOne({ _id: order.seller_id });
+    if (sellerUser === null)
+      throw new Error("sellerUser was not found");
     // We need two i18n contexts to send messages to each user
     const i18nCtxBuyer = await getUserI18nContext(buyerUser);
     const i18nCtxSeller = await getUserI18nContext(sellerUser);
@@ -135,4 +149,4 @@ const payHoldInvoice = async (bot, order) => {
   }
 };
 
-module.exports = { subscribeInvoice, payHoldInvoice };
+export { subscribeInvoice, payHoldInvoice };
