@@ -1,5 +1,4 @@
-import { generateSecretKey, getPublicKey as nostrGetPublicKey } from 'nostr-tools';
-import { SimplePool } from 'nostr-tools';
+import { generateSecretKey, getPublicKey as nostrGetPublicKey, SimplePool } from 'nostr-tools';
 
 const nostrSkEnvVar = process.env.NOSTR_SK;
 const sk = nostrSkEnvVar ? Buffer.from(nostrSkEnvVar, 'hex') : generateSecretKey();
